api/utils/direnajmongomanager_new: add refactored mongodb manager
Refactored code for direnajmongomanager.py. The old code was the collection of
functions, which imported configuration from config.py. With this approach,
tests were not easily configurable to use test database.

New manager is organized as classes and methods. It accepts mongo settings from
init which will make testing easier.

It keeps 2 connections to mongodb with pymongo and motor clients. Some parts of
the code need to be synchronous with pymongo. However, most part of the code
uses motor asynchronous client.

We keep the old mongomanager for testing purposes. When we are sure that
refactored code works, the old one will be deleted.